# Optimize lookups in FontEmbeddingManager, use HashSet over Dictionary